+++ src/components/lobby/lobby-player-card.css
@@ -1,4 +1,3 @@
-/* Lobby */
 .lobby-player-card {
   display: flex;
   padding: 1rem;
