+++ src/components/tables/lobby.table.tsx
@@ -9,62 +9,8 @@ import {
 
 import {GameState} from "types/GameState";
 
-const RoomTable = () => {
-  const data: GameState[] = [
-    {
-      hostID: "Host 1",
-      createdAt: new Date(),
-      settings: {
-        isInviteOnly: false,
-        isSpectatorAllowed: false,
-      },
-      invitedID: [],
-      status: "Waiting",
-      players: [
-        {
-          displayName: "Player 1",
-          country: "International",
-          avatarID: "2",
-          email: "hi@email.com",
-          id: "1",
-          isReady: true,
-          isHost: false,
-          numOfGamesPlayed: 0,
-          numOfGamesWon: 0,
-          position: 0,
-          roomID: "1",
-        },
-        {
-          displayName: "Player 1",
-          country: "International",
-          avatarID: "1",
-          email: "hi@email.com",
-          id: "1",
-          isReady: true,
-          isHost: false,
-          numOfGamesPlayed: 0,
-          numOfGamesWon: 0,
-          position: 0,
-          roomID: "1",
-        },
-      ],
-      biddingPhase: null,
-      trickTakingPhase: null,
-    },
-    {
-      hostID: "Host 2",
-      createdAt: new Date(),
-      settings: {
-        isInviteOnly: false,
-        isSpectatorAllowed: false,
-      },
-      invitedID: [],
-      status: "Waiting",
-      players: [],
-      biddingPhase: null,
-      trickTakingPhase: null,
-    },
-  ];
+const RoomTable = (props: {gameRoomList: GameState[]}) => {
+  const {gameRoomList} = props;
 
   const columnHelper = createColumnHelper<GameState>();
 
@@ -101,8 +47,9 @@ const RoomTable = () => {
       },
     }),
   ];
+
   const table = useReactTable({
-    data,
+    data: gameRoomList,
     columns,
     getCoreRowModel: getCoreRowModel(),
   });
